fix: store coffee prices as int cents to prevent rounding errors
prices stored as float64 could lead to rounding errors when converting to cents for stripe (19.99 * 100 = 1998.99... truncating to 1998). canged Coffee.Price to int, storing cents directly. Updated the seed data, fallback products, API handlers, and all TUI display code to be in sync with this change.

+++ pkg/models/coffee.go
@@ -13,7 +13,7 @@ type Coffee struct {
 	RoastType   string         `gorm:"size:100" json:"roast_type"`
 	Ounces      int            `json:"ounces"`
 	BeanType    string         `gorm:"size:100" json:"bean_type"`
-	Price       float64        `gorm:"not null" json:"price"`
+	Price       int            `gorm:"not null" json:"price"`
 	Color       string         `gorm:"size:7" json:"color"` // Hex color for background
 	Description string         `gorm:"type:text" json:"description"`
 	CreatedAt   time.Time      `json:"created_at"`
@@ -25,7 +25,6 @@ type Coffee struct {
 func (Coffee) TableName() string {
 	return "coffees"
 }
-
 
 // AccountMenuItems something
 var AccountMenuItems = []string{
